server/requirements.txt
supabase
python-dotenv
pyautogui
pygetwindow
pystray
pillow
plyer
pywin32
openai
pyinstaller
requests
pycaw
comtypes
pyperclip
psutil
